requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.4.0
awscli==1.42.55
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.0
boto3==1.40.55
botocore==1.40.55
bs4==0.0.2
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cssselect==1.3.0
cymem==2.0.11
docutils==0.19
fake-useragent==2.2.0
fastapi==0.119.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpx==0.27.0
idna==3.11
iniconfig==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
narwhals==2.8.0
nltk==3.9.2
numpy==2.3.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parse==1.20.2
pillow==11.3.0
playwright==1.48.0
pluggy==1.6.0
preshed==3.0.10
protobuf==6.33.0
pyarrow==21.0.0
pyasn1==0.6.1
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pyppeteer==0.0.25
pyquery==2.0.1
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-html==0.10.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.7.2
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
selenium==4.37.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.48.0
streamlit==1.50.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
w3lib==2.3.1
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
